requirements.txt
attrs==16.2.0
cffi==1.8.3
cryptography==1.5.2
cssselect==0.9.2
Django==1.10.2
djangorestframework==3.4.7
enum34==1.1.6
idna==2.1
ipaddress==1.0.17
lxml==3.6.4
parsel==1.0.3
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
PyDispatcher==2.0.5
pyOpenSSL==16.1.0
pypiwin32==219
queuelib==1.4.2
Scrapy==1.2.0
service-identity==16.0.0
six==1.10.0
Twisted==16.4.1
w3lib==1.15.0
zope.interface==4.3.2
